--- a/channel-input_test.go
+++ b/channel-input_test.go
@@ -132,48 +132,108 @@ func TestChannelInput_GetUrlsetUrl(t *testing.T) {
 }
 
 func TestWriteAll_ChannelInput(t *testing.T) {
-	RegisterTestingT(t)
+	t.Run("simple", func(t *testing.T) {
+		RegisterTestingT(t)
 
-	var out bufferOuput
-	in := NewChannelInput(func(idx int) string {
-		return fmt.Sprintf("channel input urlset %d", idx)
+		var out bufferOuput
+		in := NewChannelInput(func(idx int) string {
+			return fmt.Sprintf("channel input urlset %d", idx)
+		})
+
+		go func(in *ChannelInput) {
+			in.Feed(&UrlEntry{Loc: "a"})
+			in.Feed(&UrlEntry{Loc: "b"})
+			in.Feed(&UrlEntry{Loc: "c"})
+			in.Close()
+		}(in)
+
+		Ω(WriteAll(&out, in)).Should(BeNil())
+
+		Ω(out.index.String()).Should(MatchXML(`
+			<?xml version="1.0" encoding="UTF-8"?>
+			<sitemapindex xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
+			  <url>
+				<loc>channel input urlset 0</loc>
+			  </url>
+			</sitemapindex>
+		`))
+
+		Ω(out.sitemaps).Should(HaveLen(1))
+		Ω(out.sitemaps[0].String()).Should(MatchXML(`
+			<?xml version="1.0" encoding="UTF-8"?>
+			<urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9"
+				xmlns:image="http://www.google.com/schemas/sitemap-image/1.1">
+				<url> <loc>a</loc> </url>
+				<url> <loc>b</loc> </url>
+				<url> <loc>c</loc> </url>
+			</urlset>
+		`))
 	})
 
-	go func(in *ChannelInput) {
-		in.Feed(&UrlEntry{Loc: "a"})
-	}(in)
-	go func(in *ChannelInput) {
-		time.Sleep(100 * time.Millisecond)
-		in.Feed(&UrlEntry{Loc: "b"})
-	}(in)
-	go func(in *ChannelInput) {
-		time.Sleep(200 * time.Millisecond)
-		in.Feed(&UrlEntry{Loc: "c"})
-	}(in)
-	go func(in *ChannelInput) {
-		time.Sleep(500 * time.Millisecond)
-		in.Close()
-	}(in)
+	t.Run("concurrent", func(t *testing.T) {
+		RegisterTestingT(t)
 
-	Ω(WriteAll(&out, in)).Should(BeNil())
+		var out bufferOuput
+		in := NewChannelInput(func(idx int) string {
+			return fmt.Sprintf("channel input urlset %d", idx)
+		})
 
-	Ω(out.index.String()).Should(MatchXML(`
-<?xml version="1.0" encoding="UTF-8"?>
-<sitemapindex xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
-  <url>
-    <loc>channel input urlset 0</loc>
-  </url>
-</sitemapindex>
-	`))
+		go func(in *ChannelInput) {
+			in.Feed(&UrlEntry{Loc: "a"})
+		}(in)
+		go func(in *ChannelInput) {
+			time.Sleep(100 * time.Millisecond)
+			in.Feed(&UrlEntry{Loc: "b"})
+		}(in)
+		go func(in *ChannelInput) {
+			time.Sleep(200 * time.Millisecond)
+			in.Feed(&UrlEntry{Loc: "c"})
+		}(in)
+		go func(in *ChannelInput) {
+			time.Sleep(500 * time.Millisecond)
+			in.Close()
+		}(in)
 
-	Ω(out.sitemaps).Should(HaveLen(1))
-	Ω(out.sitemaps[0].String()).Should(MatchXML(`
-		<?xml version="1.0" encoding="UTF-8"?>
-		<urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9"
-			xmlns:image="http://www.google.com/schemas/sitemap-image/1.1">
-			<url> <loc>a</loc> </url>
-			<url> <loc>b</loc> </url>
-			<url> <loc>c</loc> </url>
-		</urlset>
-	`))
+		Ω(WriteAll(&out, in)).Should(BeNil())
+
+		Ω(out.index.String()).Should(MatchXML(`
+			<?xml version="1.0" encoding="UTF-8"?>
+			<sitemapindex xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
+			  <url>
+				<loc>channel input urlset 0</loc>
+			  </url>
+			</sitemapindex>
+		`))
+
+		Ω(out.sitemaps).Should(HaveLen(1))
+		Ω(out.sitemaps[0].String()).Should(MatchXML(`
+			<?xml version="1.0" encoding="UTF-8"?>
+			<urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9"
+				xmlns:image="http://www.google.com/schemas/sitemap-image/1.1">
+				<url> <loc>a</loc> </url>
+				<url> <loc>b</loc> </url>
+				<url> <loc>c</loc> </url>
+			</urlset>
+		`))
+	})
+
+	t.Run("multiplePages", func(t *testing.T) {
+		RegisterTestingT(t)
+
+		inputSize := 58_765
+		var out bufferOuput
+		in := NewChannelInput(func(idx int) string {
+			return fmt.Sprintf("urlset %03d", idx)
+		})
+
+		go func(in *ChannelInput) {
+			for i := 0; i < inputSize; i++ {
+				in.Feed(&UrlEntry{Loc: fmt.Sprintf("http://goiguide.com/%d", i+1)})
+			}
+			in.Close()
+		}(in)
+
+		Ω(WriteAll(&out, in)).Should(BeNil())
+		assertOutput(&out, inputSize)
+	})
 }