--- a/sitemap_test.go
+++ b/sitemap_test.go
@@ -59,7 +59,7 @@ func TestWriteAll(t *testing.T) {
 
 			totalLocs += len(s.Locs)
 			for j := 0; j < nlocs; j++ {
-				Ω(s.Locs[j]).Should(Equal(fmt.Sprintf("http://goiguide.com/%d", urlsetOffset+j)))
+				Ω(s.Locs[j]).Should(Equal(fmt.Sprintf("http://goiguide.com/%d", urlsetOffset+j+1)))
 			}
 		}
 		Ω(totalLocs).Should(Equal(expSize))
@@ -117,13 +117,6 @@ func TestWriteAll(t *testing.T) {
 <?xml version="1.0" encoding="UTF-8"?>
 <urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9" xmlns:image="http://www.google.com/schemas/sitemap-image/1.1">
   <url>
-    <loc>http://goiguide.com/0</loc>
-    <lastmod>2001-03-04T00:00:00Z</lastmod>
-    <image:image>
-      <image:loc>http://youriguide.com/0.jpg</image:loc>
-    </image:image>
-  </url>
-  <url>
     <loc>http://goiguide.com/1</loc>
     <lastmod>2001-03-04T00:00:00Z</lastmod>
     <image:image>
@@ -135,6 +128,13 @@ func TestWriteAll(t *testing.T) {
     <lastmod>2001-03-04T00:00:00Z</lastmod>
     <image:image>
       <image:loc>http://youriguide.com/2.jpg</image:loc>
+    </image:image>
+  </url>
+  <url>
+    <loc>http://goiguide.com/3</loc>
+    <lastmod>2001-03-04T00:00:00Z</lastmod>
+    <image:image>
+      <image:loc>http://youriguide.com/3.jpg</image:loc>
     </image:image>
   </url>
 </urlset>
@@ -218,18 +218,210 @@ func TestWriteAll(t *testing.T) {
 		assertOutput(&out, in.Size)
 	})
 
-	t.Run("numerousSitemaps", func(t *testing.T) {
-		RegisterTestingT(t)
+	t.Run("nextAdvances", func(t *testing.T) {
+		t.Run("inputInvariant", func(t *testing.T) {
+			RegisterTestingT(t)
 
-		in := dynamicInput{
-			Size:            50_000*11 + 97,
-			CustomEntry:     customEntry,
-			CustomUrlsetUrl: customUrl,
-		}
-		var out bufferOuput
+			in := dynamicInput{
+				Size:            3,
+				CustomEntry:     customEntry,
+				CustomUrlsetUrl: customUrl,
+			}
 
-		Ω(WriteAll(&out, &in)).Should(BeNil())
-		assertOutput(&out, in.Size)
+			Ω(in.HasNextAdvances).Should(BeFalse())
+			Ω(in.HasNext()).Should(BeTrue())
+			Ω(in.nextIdx).Should(Equal(0))
+			Ω(in.HasNext()).Should(BeTrue())
+			Ω(in.nextIdx).Should(Equal(0))
+
+			Ω(in.Next()).ShouldNot(BeNil())
+			Ω(in.nextIdx).Should(Equal(1))
+			Ω(in.HasNext()).Should(BeTrue())
+			Ω(in.nextIdx).Should(Equal(1))
+			Ω(in.HasNext()).Should(BeTrue())
+			Ω(in.nextIdx).Should(Equal(1))
+		})
+
+		t.Run("shortSingleSitemap", func(t *testing.T) {
+			RegisterTestingT(t)
+
+			in := dynamicInput{
+				Size:            3,
+				CustomEntry:     customEntry,
+				CustomUrlsetUrl: customUrl,
+			}
+			var out bufferOuput
+
+			Ω(in.HasNextAdvances).Should(BeFalse())
+			Ω(WriteAll(&out, &in)).Should(BeNil())
+			assertOutput(&out, in.Size)
+		})
+
+		t.Run("maxSingleSitemap", func(t *testing.T) {
+			RegisterTestingT(t)
+
+			in := dynamicInput{
+				Size:            50_000,
+				CustomEntry:     customEntry,
+				CustomUrlsetUrl: customUrl,
+			}
+			var out bufferOuput
+
+			Ω(in.HasNextAdvances).Should(BeFalse())
+			Ω(WriteAll(&out, &in)).Should(BeNil())
+			assertOutput(&out, in.Size)
+		})
+
+		t.Run("minTwoSitemaps", func(t *testing.T) {
+			RegisterTestingT(t)
+
+			in := dynamicInput{
+				Size:            50_000 + 1,
+				CustomEntry:     customEntry,
+				CustomUrlsetUrl: customUrl,
+			}
+			var out bufferOuput
+
+			Ω(in.HasNextAdvances).Should(BeFalse())
+			Ω(WriteAll(&out, &in)).Should(BeNil())
+			assertOutput(&out, in.Size)
+		})
+
+		t.Run("maxTwoSitemaps", func(t *testing.T) {
+			RegisterTestingT(t)
+
+			in := dynamicInput{
+				Size:            50_000 * 2,
+				CustomEntry:     customEntry,
+				CustomUrlsetUrl: customUrl,
+			}
+			var out bufferOuput
+
+			Ω(in.HasNextAdvances).Should(BeFalse())
+			Ω(WriteAll(&out, &in)).Should(BeNil())
+			assertOutput(&out, in.Size)
+		})
+
+		t.Run("multipleSitemaps", func(t *testing.T) {
+			RegisterTestingT(t)
+
+			in := dynamicInput{
+				Size:            50_000*3 + 123,
+				CustomEntry:     customEntry,
+				CustomUrlsetUrl: customUrl,
+			}
+			var out bufferOuput
+
+			Ω(in.HasNextAdvances).Should(BeFalse())
+			Ω(WriteAll(&out, &in)).Should(BeNil())
+			assertOutput(&out, in.Size)
+		})
+	})
+
+	t.Run("hasNextAdvances", func(t *testing.T) {
+		t.Run("inputInvariant", func(t *testing.T) {
+			RegisterTestingT(t)
+
+			in := dynamicInput{
+				Size:            3,
+				CustomEntry:     customEntry,
+				CustomUrlsetUrl: customUrl,
+				HasNextAdvances: true,
+			}
+
+			Ω(in.HasNextAdvances).Should(BeTrue())
+			Ω(in.Next()).ShouldNot(BeNil())
+			Ω(in.nextIdx).Should(Equal(0))
+			Ω(in.Next()).ShouldNot(BeNil())
+			Ω(in.nextIdx).Should(Equal(0))
+
+			Ω(in.HasNext()).Should(BeTrue())
+			Ω(in.nextIdx).Should(Equal(1))
+			Ω(in.Next()).ShouldNot(BeNil())
+			Ω(in.nextIdx).Should(Equal(1))
+			Ω(in.Next()).ShouldNot(BeNil())
+			Ω(in.nextIdx).Should(Equal(1))
+		})
+
+		t.Run("shortSingleSitemap", func(t *testing.T) {
+			RegisterTestingT(t)
+
+			in := dynamicInput{
+				Size:            3,
+				CustomEntry:     customEntry,
+				CustomUrlsetUrl: customUrl,
+				HasNextAdvances: true,
+			}
+			var out bufferOuput
+
+			Ω(in.HasNextAdvances).Should(BeTrue())
+			Ω(WriteAll(&out, &in)).Should(BeNil())
+			assertOutput(&out, in.Size)
+		})
+
+		t.Run("maxSingleSitemap", func(t *testing.T) {
+			RegisterTestingT(t)
+
+			in := dynamicInput{
+				Size:            50_000,
+				CustomEntry:     customEntry,
+				CustomUrlsetUrl: customUrl,
+				HasNextAdvances: true,
+			}
+			var out bufferOuput
+
+			Ω(in.HasNextAdvances).Should(BeTrue())
+			Ω(WriteAll(&out, &in)).Should(BeNil())
+			assertOutput(&out, in.Size)
+		})
+
+		t.Run("minTwoSitemaps", func(t *testing.T) {
+			RegisterTestingT(t)
+
+			in := dynamicInput{
+				Size:            50_000 + 1,
+				CustomEntry:     customEntry,
+				CustomUrlsetUrl: customUrl,
+				HasNextAdvances: true,
+			}
+			var out bufferOuput
+
+			Ω(in.HasNextAdvances).Should(BeTrue())
+			Ω(WriteAll(&out, &in)).Should(BeNil())
+			assertOutput(&out, in.Size)
+		})
+
+		t.Run("maxTwoSitemaps", func(t *testing.T) {
+			RegisterTestingT(t)
+
+			in := dynamicInput{
+				Size:            50_000 * 2,
+				CustomEntry:     customEntry,
+				CustomUrlsetUrl: customUrl,
+				HasNextAdvances: true,
+			}
+			var out bufferOuput
+
+			Ω(in.HasNextAdvances).Should(BeTrue())
+			Ω(WriteAll(&out, &in)).Should(BeNil())
+			assertOutput(&out, in.Size)
+		})
+
+		t.Run("multipleSitemaps", func(t *testing.T) {
+			RegisterTestingT(t)
+
+			in := dynamicInput{
+				Size:            50_000*3 + 123,
+				CustomEntry:     customEntry,
+				CustomUrlsetUrl: customUrl,
+				HasNextAdvances: true,
+			}
+			var out bufferOuput
+
+			Ω(in.HasNextAdvances).Should(BeTrue())
+			Ω(WriteAll(&out, &in)).Should(BeNil())
+			assertOutput(&out, in.Size)
+		})
 	})
 
 	t.Run("failures", func(t *testing.T) {
@@ -696,6 +888,7 @@ type dynamicInput struct {
 	DefaultEntry    UrlEntry
 	CustomEntry     func(int) *UrlEntry
 	CustomUrlsetUrl func(int) string
+	HasNextAdvances bool
 
 	nextIdx int
 }
@@ -704,13 +897,19 @@ func (d *dynamicInput) Reset() {
 	d.nextIdx = 0
 }
 
-func (d dynamicInput) HasNext() bool {
-	return d.nextIdx < d.Size
+func (d *dynamicInput) HasNext() bool {
+	idx := d.nextIdx
+	if d.HasNextAdvances {
+		d.nextIdx++
+	}
+	return idx < d.Size
 }
 
 func (d *dynamicInput) Next() *UrlEntry {
+	if !d.HasNextAdvances {
+		d.nextIdx++
+	}
 	idx := d.nextIdx
-	d.nextIdx++
 	if d.CustomEntry != nil {
 		return d.CustomEntry(idx)
 	}